# docs/contracts/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Educational Framework Backend API",
    "description": "Family module API contract.",
    "version": "0.0.1"
  },
...
  "paths": {
...
    "/api/v1/dev/content/topics": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create topic",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTopicRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseTopic" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found (category not found)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } },
          "409": { "description": "Conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error409" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List topics, optionally filtered by category",
        "parameters": [{ "name": "categoryId", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseTopicList" } } } }
        }
      }
    },
    "/api/v1/dev/content/topics/{id}": {
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Get topic by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseTopic" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      },
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update topic",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTopicRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseTopic" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
    "/api/v1/dev/content/difficulty-levels": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create difficulty level for an activity",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateDifficultyLevelRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseDifficultyLevel" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found (activity not found)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List difficulty levels for an activity",
        "parameters": [{ "name": "activityId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseDifficultyLevelList" } } } }
        }
      }
    },
    "/api/v1/dev/content/difficulty-levels/{id}": {
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update difficulty level",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateDifficultyLevelRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseDifficultyLevel" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
    "/api/v1/dev/content/locales": {
      "post": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Create content locale",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateContentLocaleRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseContentLocale" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "409": { "description": "Conflict (locale already exists for entity)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error409" } } } }
        }
      },
      "get": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] List locales for an entity",
        "parameters": [
          { "name": "entityType", "in": "query", "required": true, "schema": { "type": "string", "enum": ["CATEGORY", "TOPIC", "ACTIVITY", "DIFFICULTY_LEVEL", "ACTIVITY_RESOURCE"] } },
          { "name": "entityId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseContentLocaleList" } } } }
        }
      }
    },
    "/api/v1/dev/content/locales/{id}": {
      "put": {
        "tags": ["dev-content"],
        "summary": "[DEV ONLY] Update content locale",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateContentLocaleRequest" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponseContentLocale" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } } },
          "404": { "description": "Not Found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque family session token returned only by POST /api/v1/auth/login."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "data": { "nullable": true },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateTopicRequest": {
        "type": "object",
        "required": ["name", "categoryId", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "categoryId": { "type": "integer", "format": "int64" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "compatibleVariants": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "UpdateTopicRequest": {
        "type": "object",
        "required": ["name", "categoryId", "status"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true },
          "categoryId": { "type": "integer", "format": "int64" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "compatibleVariants": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "TopicResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "categoryId": { "type": "integer", "format": "int64" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "DRAFT"] },
          "minAge": { "type": "integer", "nullable": true },
          "maxAge": { "type": "integer", "nullable": true },
          "compatibleVariants": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseTopic": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/TopicResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseTopicList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/TopicResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateDifficultyLevelRequest": {
        "type": "object",
        "required": ["activityId", "difficultyCode"],
        "properties": {
          "activityId": { "type": "integer", "format": "int64" },
          "difficultyCode": { "type": "string", "enum": ["EASY", "MEDIUM", "HARD"] },
          "engineParams": { "type": "string", "nullable": true, "description": "JSON string with engine-specific parameters." },
          "adaptiveThresholdConfig": { "type": "string", "nullable": true, "description": "JSON string with adaptive threshold configuration." }
        }
      },
      "UpdateDifficultyLevelRequest": {
        "type": "object",
        "required": ["difficultyCode"],
        "properties": {
          "difficultyCode": { "type": "string", "enum": ["EASY", "MEDIUM", "HARD"] },
          "engineParams": { "type": "string", "nullable": true },
          "adaptiveThresholdConfig": { "type": "string", "nullable": true }
        }
      },
      "DifficultyLevelResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "activityId": { "type": "integer", "format": "int64" },
          "difficultyCode": { "type": "string", "enum": ["EASY", "MEDIUM", "HARD"] },
          "engineParams": { "type": "string", "nullable": true },
          "adaptiveThresholdConfig": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseDifficultyLevel": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/DifficultyLevelResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseDifficultyLevelList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/DifficultyLevelResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
...
      "CreateContentLocaleRequest": {
        "type": "object",
        "required": ["entityType", "entityId", "localeCode", "name"],
        "properties": {
          "entityType": { "type": "string", "enum": ["CATEGORY", "TOPIC", "ACTIVITY", "DIFFICULTY_LEVEL", "ACTIVITY_RESOURCE"] },
          "entityId": { "type": "integer", "format": "int64" },
          "localeCode": { "type": "string", "example": "es-ES", "description": "Only es-ES is supported in v1." },
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true }
        }
      },
      "UpdateContentLocaleRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "nullable": true }
        }
      },
      "ContentLocaleResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "entityType": { "type": "string" },
          "entityId": { "type": "integer", "format": "int64" },
          "localeCode": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiResponseContentLocale": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ContentLocaleResponse" },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "ApiResponseContentLocaleList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ContentLocaleResponse" } },
          "message": { "type": "string", "nullable": true },
          "errors": { "type": "array", "nullable": true, "items": { "type": "string" } }
        }
      },
      "Error400": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Bad Request error response."
      },
...
      "Error404": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Not Found error response."
      },
      "Error409": {
        "allOf": [{ "$ref": "#/components/schemas/ApiResponse" }],
        "description": "Conflict error response."
      },
...
    }
  }
}