# framework/backend/src/main/java/es/vargontoc/educational/framework/content/infrastructure/dto/CreateContentLocaleRequest.java
package es.vargontoc.educational.framework.content.infrastructure.dto;

public record CreateContentLocaleRequest(
    String entityType,
    Long entityId,
    String localeCode,
    String name,
    String description
) {}