# framework/backend/src/main/java/es/vargontoc/educational/framework/content/infrastructure/dto/UpdateDifficultyLevelRequest.java
package es.vargontoc.educational.framework.content.infrastructure.dto;

public record UpdateDifficultyLevelRequest(
    String difficultyCode,
    String engineParams,
    String adaptiveThresholdConfig
) {}